add util/time, use for lvgl port
