+++ src/util/time.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <stdint.h>
+
+uint32_t time_ms();
+uint32_t time_s();
